// packages/esm-patient-vitals-app/src/vitals-biometrics-form/types.ts
export interface VitalsBiometricsFormData {
  systolicBloodPressure: number;
  diastolicBloodPressure: number;
  respiratoryRate: number;
  oxygenSaturation: number;
  pulse: number;
  temperature: number;
  generalPatientNote: string;
  weight: number;
  height: number;
  midUpperArmCircumference: number;
  computedBodyMassIndex: number;
}